moved mock struct and implementation methods to cmd runner file and refactored all go command files to proper naming convention.
