+++ Change_Logs.md
@@ -368,3 +368,4 @@ _**NOTE**_ : *`put`* method refactored to *`create`*
        Flags:
          -h, --help   help for update
     ```
+
