--- cmd/cmd_runner.go
@@ -1,9 +1,0 @@
-package cmd
-
-import (
-	"CacheCLI/kvcache"
-)
-
-type CommandRunner struct {
-	cache kvcache.KeyValueCache
-}
