+++ cmd/root.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"KVCache/kvcache"
-	"errors"
 	"fmt"
 	"github.com/spf13/cobra"
 	"os"
@@ -45,23 +44,7 @@ var deleteCmd = &cobra.Command{
 	Args: cobra.MinimumNArgs(1),
 	Short: "delete key-value pair",
 	Long:  "delete key value strings into the key-value cache",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		if cache == nil {
-			return errors.New("cache not initialized - delete failed: ")
-		}
-		//pre-seeding cache for read command for now since cache won't persist until CLI/Cache connection built
-		cache.Create("name", "harley")
-		cache.Create("animal", "horse")
-		cache.Create("kitten", "Bene")
-		deleteResult := cache.Delete(args[0])
-		if deleteResult == nil {
-			fmt.Printf("delete success: cache '%v' ", cache)
-			fmt.Println()
-			return nil
-		}
-		fmt.Println(deleteResult)
-		return errors.New("")
-	},
+	//RunE:
 }
